+++ projects/app/src/app/features/inventory/constants/confirm-prompt.modal.ts
@@ -1,13 +1,6 @@
 import { ConfirmPromptModalInput } from '@app/common/components';
-import * as listMenu from '../contextual-menus/list';
 import * as categoryMenu from '../contextual-menus/category';
 import * as itemMenu from '../contextual-menus/item';
-
-// export const LIST_REMOVE_PROMPT: ConfirmPromptModalInput = {
-//   action: listMenu.LIST_ACTION_REMOVE.id,
-//   title: 'common.prompt.removeAll.title',
-//   message: 'common.prompt.removeAll.message',
-// };
 
 export const CATEGORY_REMOVE_PROMPT: ConfirmPromptModalInput = {
   action: categoryMenu.CATEGORY_ACTION_REMOVE.id,
