+++ projects/app/src/app/features/list/store/category.ts
@@ -1,27 +1,25 @@
 import { Subscription } from 'rxjs';
 
 import { DEFAULT_CATEGORY } from '@app/core/constants';
-import { updateStore } from '@app/common/store';
+import { updateStore } from '@common/store';
 import { InventoryItem } from '@app/features/inventory';
 import { ListStore } from './feature';
 import { CreateListItemDto } from '../types';
 
 export class ListCategoryItemsSubstore {
-
-  constructor(
-    private parent: ListStore,
-  ) {}
+  constructor(private parent: ListStore) {}
 
   cloneFromInventory(inventoryItems: InventoryItem[]): Subscription {
-
     const itemNamesSet = new Set<string>();
-    this.parent.items().forEach(item => itemNamesSet.add(item.name.toLowerCase()));
+    this.parent
+      .items()
+      .forEach((item) => itemNamesSet.add(item.name.toLowerCase()));
 
     // Pick only the inventory items that do not yet exist in the List
     // Map the inventory items to a list of CreateListItemDto types
     const listItems: CreateListItemDto[] = inventoryItems
-      .filter(item => !itemNamesSet.has(item.name.toLowerCase()))
-      .map(inventoryItem => ({
+      .filter((item) => !itemNamesSet.has(item.name.toLowerCase()))
+      .map((inventoryItem) => ({
         category: inventoryItem?.category ?? '',
         name: inventoryItem.name,
         amount: 1,
@@ -32,17 +30,18 @@ export class ListCategoryItemsSubstore {
   }
 
   createMany(listItems: CreateListItemDto[]): Subscription {
-
-    const category = listItems.length ? listItems[0].category : DEFAULT_CATEGORY;
+    const category = listItems.length
+      ? listItems[0].category
+      : DEFAULT_CATEGORY;
 
     return updateStore(this.parent.api.categoryItems.createMany(listItems))
       .withFeedback(this.parent.feedback)
       .withNotifications(
         ['inventory.cloneCategoryToList.success', { categoryName: category }],
-        ['inventory.cloneCategoryToList.error', { categoryName: category }],
+        ['inventory.cloneCategoryToList.error', { categoryName: category }]
       )
-      .onSuccess(newItemsInCategory => {
-        this.parent.items.update(items => [...items, ...newItemsInCategory]);
+      .onSuccess((newItemsInCategory) => {
+        this.parent.items.update((items) => [...items, ...newItemsInCategory]);
       })
       .update();
   }
@@ -52,13 +51,15 @@ export class ListCategoryItemsSubstore {
       .withFeedback(this.parent.feedback)
       .withNotifications(
         'common.async.editItemsSuccess',
-        'common.async.editItemsError',
+        'common.async.editItemsError'
       )
       .onSuccess(() => {
-        this.parent.items.update(prev => prev.map(item => {
-          if (item.category !== category) return item;
-          return { ...item, isDone: true };
-        }));
+        this.parent.items.update((prev) =>
+          prev.map((item) => {
+            if (item.category !== category) return item;
+            return { ...item, isDone: true };
+          })
+        );
       })
       .update();
   }
@@ -68,13 +69,15 @@ export class ListCategoryItemsSubstore {
       .withFeedback(this.parent.feedback)
       .withNotifications(
         'common.async.editItemsSuccess',
-        'common.async.editItemsError',
+        'common.async.editItemsError'
       )
       .onSuccess(() => {
-        this.parent.items.update(prev => prev.map(item => {
-          if (item.category !== category) return item;
-          return { ...item, isDone: false };
-        }));
+        this.parent.items.update((prev) =>
+          prev.map((item) => {
+            if (item.category !== category) return item;
+            return { ...item, isDone: false };
+          })
+        );
       })
       .update();
   }
@@ -84,13 +87,15 @@ export class ListCategoryItemsSubstore {
       .withFeedback(this.parent.feedback)
       .withNotifications(
         'common.async.removeItemsSuccess',
-        'common.async.removeItemsError',
+        'common.async.removeItemsError'
       )
       .onSuccess(() => {
-        this.parent.items.update(prev => prev.filter(item => {
-          if (item.category !== category) return true;
-          return false;
-        }));
+        this.parent.items.update((prev) =>
+          prev.filter((item) => {
+            if (item.category !== category) return true;
+            return false;
+          })
+        );
       })
       .update();
   }
@@ -100,13 +105,15 @@ export class ListCategoryItemsSubstore {
       .withFeedback(this.parent.feedback)
       .withNotifications(
         'common.async.removeItemsSuccess',
-        'common.async.removeItemsError',
+        'common.async.removeItemsError'
       )
       .onSuccess(() => {
-        this.parent.items.update(prev => prev.filter(item => {
-          if (item.category !== category) return true;
-          return !item.isDone;
-        }));
+        this.parent.items.update((prev) =>
+          prev.filter((item) => {
+            if (item.category !== category) return true;
+            return !item.isDone;
+          })
+        );
       })
       .update();
   }
