+++ projects/app/src/app/features/list/store/item.ts
@@ -1,23 +1,20 @@
-import { removeItem, updateStore, updateItem } from '@app/common/store';
+import { removeItem, updateStore, updateItem } from '@common/store';
 import { CreateListItemDto, ListItem } from '../types';
 import { ListStore } from './feature';
 
 export class ListItemSubstore {
-
-  constructor(
-    private parent: ListStore,
-  ) {}
+  constructor(private parent: ListStore) {}
 
   create(dto: CreateListItemDto) {
     return updateStore(this.parent.api.item.create(dto))
       .withFeedback(this.parent.feedback)
       .withNotifications(
         'common.async.createItemSuccess',
-        'common.async.createItemError',
+        'common.async.createItemError'
       )
-      .onSuccess(item => {
-        this.parent.itemModalSuccessCounter.update(counter => counter + 1);
-        this.parent.items.update(items => [...items, item]);
+      .onSuccess((item) => {
+        this.parent.itemModalSuccessCounter.update((counter) => counter + 1);
+        this.parent.items.update((items) => [...items, item]);
       })
       .update();
   }
@@ -27,10 +24,10 @@ export class ListItemSubstore {
       .withFeedback(this.parent.feedback)
       .withNotifications(
         'common.async.editItemSuccess',
-        'common.async.editItemError',
+        'common.async.editItemError'
       )
-      .onSuccess(updated => {
-        this.parent.itemModalSuccessCounter.update(counter => counter + 1);
+      .onSuccess((updated) => {
+        this.parent.itemModalSuccessCounter.update((counter) => counter + 1);
         updateItem(this.parent.items, updated.id, updated);
       })
       .update();
@@ -39,7 +36,7 @@ export class ListItemSubstore {
   complete(itemId: ListItem['id']) {
     return updateStore(this.parent.api.item.complete(itemId))
       .withFeedback(this.parent.feedback)
-      .onSuccess(updated => {
+      .onSuccess((updated) => {
         updateItem(this.parent.items, updated.id, updated);
       })
       .update();
@@ -48,7 +45,7 @@ export class ListItemSubstore {
   undo(itemId: ListItem['id']) {
     return updateStore(this.parent.api.item.undo(itemId))
       .withFeedback(this.parent.feedback)
-      .onSuccess(updated => {
+      .onSuccess((updated) => {
         updateItem(this.parent.items, updated.id, updated);
       })
       .update();
@@ -57,7 +54,7 @@ export class ListItemSubstore {
   toggle(itemId: ListItem['id']) {
     return updateStore(this.parent.api.item.toggle(itemId))
       .withFeedback(this.parent.feedback)
-      .onSuccess(updated => {
+      .onSuccess((updated) => {
         updateItem(this.parent.items, updated.id, updated);
       })
       .update();
@@ -66,7 +63,7 @@ export class ListItemSubstore {
   increment(itemId: ListItem['id']) {
     return updateStore(this.parent.api.item.increment(itemId))
       .withFeedback(this.parent.feedback)
-      .onSuccess(updated => {
+      .onSuccess((updated) => {
         updateItem(this.parent.items, updated.id, updated);
       })
       .update();
@@ -75,7 +72,7 @@ export class ListItemSubstore {
   decrement(itemId: ListItem['id']) {
     return updateStore(this.parent.api.item.decrement(itemId))
       .withFeedback(this.parent.feedback)
-      .onSuccess(updated => {
+      .onSuccess((updated) => {
         updateItem(this.parent.items, updated.id, updated);
       })
       .update();
@@ -84,7 +81,7 @@ export class ListItemSubstore {
   remove(itemId: ListItem['id']) {
     return updateStore(this.parent.api.item.remove(itemId))
       .withFeedback(this.parent.feedback)
-      .onSuccess(removed => {
+      .onSuccess((removed) => {
         removeItem(this.parent.items, removed.id);
       })
       .update();
