+++ projects/common/src/rxjs/public-api.ts
@@ -1,6 +1,2 @@
 export * from './filter-null';
-export * from './first-false';
-export * from './first-having';
-export * from './first-non-null';
 export * from './first-truthy';
-export * from './promise-to-observable';
