+++ projects/common/src/store/collection/should-fetch.ts
@@ -1,0 +1,15 @@
+import { LOADING_STATUS, LoadingStatus, UnixTimestamp } from '@common/types';
+
+export function shouldFetchCollection<T extends Record<string, any>>(
+  items: T[],
+  status: LoadingStatus,
+  lastUpdated: UnixTimestamp | null,
+  cacheMaxAge: number = 30_000 // 30 seconds
+) {
+  return (
+    status === LOADING_STATUS.PRISTINE ||
+    lastUpdated === null ||
+    Date.now() - lastUpdated > cacheMaxAge ||
+    items.length === 0
+  );
+}
