+++ projects/common/src/store/feedback.ts
@@ -1,7 +1,7 @@
 import { WritableSignal } from '@angular/core';
-import { StoreFeedback } from '@app/common/store';
 import { LOADING_STATUS, LoadingStatus } from '@common/types';
 import { HashMap } from '@jsverse/transloco';
+import { StoreFeedback } from '@common/store';
 
 import { UiStore } from '@app/core/ui/store';
