+++ projects/common/src/store/public-api.ts
@@ -1,4 +1,10 @@
-export * from './collection/collection';
+export * from './collection/count';
+export * from './collection/filter';
+export * from './collection/group';
+export * from './collection/search';
+export * from './collection/should-fetch';
+export * from './collection/sort';
+export * from './collection/types';
 export * from './feedback';
 export * from './filters';
 export * from './update-store';
